Refactor API upload tests into their own test class.
